third_party/blink/web_tests/html/popup/popup-light-dismiss-keyboard.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>Popover light dismiss behavior</title>
<link rel="author" href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popup.research.explainer">
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>

<!-- This should be a WPT, but the test_driver.Actions()
     system doesn't properly do keyDown/keyUp.
     When that situation has changed, this should be
     re-included with this WPT test:
     external/wpt/html/semantics/popovers/popover-light-dismiss.tentative.html
-->

<button id=b1 onclick='p1.show()'>Popover 1</button>
<div popover id=p1 anchor=b1>Inside popover 1
  <button id=b2 onclick='p2.show()'>Popover 2</button>
</div>
<div popover id=p2 anchor=b2>Inside popover 2</div>

<style>
  #p1 { top:50px; }
  #p2 { top:50px; left:250px; }
  [popover] { border: 5px solid red; }
</style>


<script>
  test(() => {
    if (!window.eventSender)
      assert_true(false, 'eventSender is required for this test');

    const popover1 = document.querySelector('#p1');
    const popover2 = document.querySelector('#p2');
    popover1.showPopover();
    popover2.showPopover();
    eventSender.keyDown("Escape");
    assert_true(popover1.matches(':open'));
    assert_false(popover2.matches(':open'));
    eventSender.keyDown("Escape");
    assert_false(popover1.matches(':open'));
    assert_false(popover2.matches(':open'));
  },'Escape key should close top level popover');
</script>
